Created some fem toolbox classes
