+++ Common/src/CConfig.cpp
@@ -27,12 +27,13 @@
 
 #define ENABLE_MAPS
 #include <utility>
+#include <limits>
 
 #include "../include/CConfig.hpp"
 #undef ENABLE_MAPS
 
 #include "../include/fem/fem_gauss_jacobi_quadrature.hpp"
-#include "../include/fem/fem_geometry_structure.hpp"
+#include "../include/toolboxes/classes_multiple_integers.hpp"
 
 #include "../include/basic_types/ad_structure.hpp"
 #include "../include/toolboxes/printing_toolbox.hpp"
