+++ Common/src/fem/fem_geometry_structure.cpp
@@ -26,6 +26,10 @@
  */
 
 #include "../../include/toolboxes/CLinearPartitioner.hpp"
+#include "../../include/toolboxes/classes_multiple_integers.hpp"
+#include "../../include/toolboxes/fem/CReorderElements.hpp"
+#include "../../include/toolboxes/fem/CSortFaces.hpp"
+#include "../../include/toolboxes/fem/CSortBoundaryFaces.hpp"
 #include "../../include/fem/fem_geometry_structure.hpp"
 #include "../../include/geometry/primal_grid/CPrimalGridFEM.hpp"
 #include "../../include/geometry/primal_grid/CPrimalGridBoundFEM.hpp"
@@ -37,140 +41,6 @@
 extern "C" void dpotrf_(char*, int*, passivedouble*, int*, int*);
 extern "C" void dpotri_(char*, int*, passivedouble*, int*, int*);
 #endif
-
-bool CLong3T::operator<(const CLong3T& other) const {
-  if (long0 != other.long0) return (long0 < other.long0);
-  if (long1 != other.long1) return (long1 < other.long1);
-  if (long2 != other.long2) return (long2 < other.long2);
-
-  return false;
-}
-
-CReorderElements::CReorderElements(const unsigned long val_GlobalElemID, const unsigned short val_TimeLevel,
-                                   const bool val_CommSolution, const unsigned short val_VTK_Type,
-                                   const unsigned short val_nPolySol, const bool val_JacConstant) {
-  /* Copy the global elment ID, time level and whether or not this element
-     must be communicated. */
-  globalElemID = val_GlobalElemID;
-  timeLevel = val_TimeLevel;
-  commSolution = val_CommSolution;
-
-  /* Create the element type used in this class, which stores information of
-     the VTK type, the polynomial degree of the solution and whether or not the
-     Jacobian of the transformation is constant. As it is possible that the
-     polynomial degree of the solution is zero, this convention is different
-     from the convention used in the SU2 grid file. */
-  elemType = val_VTK_Type + 100 * val_nPolySol;
-  if (!val_JacConstant) elemType += 50;
-}
-
-bool CReorderElements::operator<(const CReorderElements& other) const {
-  /* Elements with the lowest time level are stored first. */
-  if (timeLevel != other.timeLevel) return timeLevel < other.timeLevel;
-
-  /* Next comparison is whether or not the element must communicate its
-     solution data to other ranks. Elements which do not need to do this
-     are stored first. */
-  if (commSolution != other.commSolution) return other.commSolution;
-
-  /* Elements of the same element type must be stored as contiguously as
-     possible to allow for the simultaneous treatment of elements in the
-     matrix multiplications. */
-  if (elemType != other.elemType) return elemType < other.elemType;
-
-  /* The final comparison is based on the global element ID. */
-  return globalElemID < other.globalElemID;
-}
-
-bool CSortFaces::operator()(const CFaceOfElement& f0, const CFaceOfElement& f1) {
-  /*--- Comparison in case both faces are boundary faces. ---*/
-  if (f0.faceIndicator >= 0 && f1.faceIndicator >= 0) {
-    /* Both faces are boundary faces. The first comparison is the boundary
-       marker, which is stored in faceIndicator. */
-    if (f0.faceIndicator != f1.faceIndicator) return f0.faceIndicator < f1.faceIndicator;
-
-    /* Both faces belong to the same boundary marker. The second comparison is
-       based on the on the local volume ID's of the adjacent elements. As the
-       volumes are sorted according to the time levels for time accurate
-       local stepping, there is no need to do this check seperately here. */
-    unsigned long ind0 = f0.elemID0 < nVolElemTot ? f0.elemID0 : f0.elemID1;
-    unsigned long ind1 = f1.elemID0 < nVolElemTot ? f1.elemID0 : f1.elemID1;
-
-    return ind0 < ind1;
-  }
-
-  /*--- Comparison in case both faces are internal faces. ---*/
-  if (f0.faceIndicator == -1 && f1.faceIndicator == -1) {
-    /* Both faces are internal faces. First determine the minimum and maximum
-       ID of its adjacent elements.  */
-    unsigned long elemIDMin0 = min(f0.elemID0, f0.elemID1);
-    unsigned long elemIDMax0 = max(f0.elemID0, f0.elemID1);
-
-    unsigned long elemIDMin1 = min(f1.elemID0, f1.elemID1);
-    unsigned long elemIDMax1 = max(f1.elemID0, f1.elemID1);
-
-    /* Determine the situation. */
-    if (elemIDMax0 < nVolElemTot && elemIDMax1 < nVolElemTot) {
-      /* Both faces are matching internal faces. Determine whether or not these
-         faces are local faces, i.e. faces between locally owned elements. */
-      const bool face0IsLocal = elemIDMax0 < nVolElemOwned;
-      const bool face1IsLocal = elemIDMax1 < nVolElemOwned;
-
-      /* Check if both faces have the same status, i.e. either local or
-         not local. */
-      if (face0IsLocal == face1IsLocal) {
-        /* Both faces are either local or not local. Determine the time level
-           of the faces, which is the minimum value of the adjacent volume
-           elements. */
-        const unsigned short timeLevel0 = min(volElem[elemIDMin0].timeLevel, volElem[elemIDMax0].timeLevel);
-        const unsigned short timeLevel1 = min(volElem[elemIDMin1].timeLevel, volElem[elemIDMax1].timeLevel);
-
-        /* Internal faces with the same status are first sorted according to
-           their time level. Faces with the smallest time level are numbered
-           first. Note this is only relevant for time accurate local time
-           stepping. */
-        if (timeLevel0 != timeLevel1) return timeLevel0 < timeLevel1;
-
-        /* The faces belong to the same time level. They are sorted according
-           to their element ID's in order to increase cache performance. */
-        if (elemIDMin0 != elemIDMin1) return elemIDMin0 < elemIDMin1;
-        return elemIDMax0 < elemIDMax1;
-      } /* One face is a local face and the other is not. Make sure that
-   the local faces are numbered first. */
-      return face0IsLocal;
-
-    } else if (elemIDMax0 >= nVolElemTot && elemIDMax1 >= nVolElemTot) {
-      /* Both faces are non-matching internal faces. Sort them according to
-         their relevant element ID. The time level is not taken into account
-         yet, because non-matching faces are not possible at the moment with
-         time accurate local time stepping. */
-      return elemIDMin0 < elemIDMin1;
-    } else {
-      /* One face is a matching internal face and the other face is a
-         non-matching internal face. Make sure that the non-matching face
-         is numbered after the matching face. This is accomplished by comparing
-         the maximum element ID's. */
-      return elemIDMax0 < elemIDMax1;
-    }
-  }
-
-  /*--- One face is a boundary face and the other face is an internal face.
-        Make sure that the boundary face is numbered first. This can be
-        accomplished by using the > operator for faceIndicator. ---*/
-  return f0.faceIndicator > f1.faceIndicator;
-}
-
-bool CSortBoundaryFaces::operator()(const CSurfaceElementFEM& f0, const CSurfaceElementFEM& f1) {
-  /* First sorting criterion is the index of the standard element. The
-     boundary faces should be sorted per standard element. Note that the
-     time level is not taken into account here, because it is assumed that
-     the surface elements to be sorted belong to one time level. */
-  if (f0.indStandardElement != f1.indStandardElement) return f0.indStandardElement < f1.indStandardElement;
-
-  /* The standard elements are the same. The second criterion is the
-     corresponding volume IDs of the surface elements. */
-  return f0.volElemID < f1.volElemID;
-}
 
 bool CPointFEM::operator<(const CPointFEM& other) const {
   if (periodIndexToDonor != other.periodIndexToDonor) return periodIndexToDonor < other.periodIndexToDonor;
