+++ Common/src/geometry/CPhysicalGeometryFEM.cpp
@@ -1,6 +1,6 @@
 /*!
- * \file geometry_structure_fem_part.cpp
- * \brief Main subroutines for distributin the grid for the Fluid FEM solver.
+ * \file CPhysicalGeometryFEM.cpp
+ * \brief Implementation of the FEM related functions of CPhysicalGeometry.
  * \author F. Palacios, T. Economon
  * \version 8.1.0 "Harrier"
  *
@@ -26,6 +26,9 @@
  */
 
 #include "../../include/toolboxes/CLinearPartitioner.hpp"
+#include "../../include/toolboxes/classes_multiple_integers.hpp"
+#include "../../include/toolboxes/fem/CFaceOfElement.hpp"
+#include "../../include/toolboxes/fem/CMatchingFace.hpp"
 #include "../../include/geometry/CPhysicalGeometry.hpp"
 #include "../../include/fem/fem_standard_element.hpp"
 #include "../../include/geometry/primal_grid/CPrimalGridFEM.hpp"
@@ -37,329 +40,6 @@
 #include <iomanip>
 #include <sys/types.h>
 #include <sys/stat.h>
-/*--- Epsilon definition ---*/
-
-CFaceOfElement::CFaceOfElement() {
-  nCornerPoints = 0;
-  cornerPoints[0] = cornerPoints[1] = cornerPoints[2] = cornerPoints[3] = ULONG_MAX;
-  elemID0 = elemID1 = ULONG_MAX;
-  nPolyGrid0 = nPolyGrid1 = 0;
-  nPolySol0 = nPolySol1 = 0;
-  nDOFsElem0 = nDOFsElem1 = 0;
-  elemType0 = elemType1 = 0;
-  faceID0 = faceID1 = 0;
-  periodicIndex = periodicIndexDonor = 0;
-  faceIndicator = 0;
-
-  JacFaceIsConsideredConstant = false;
-  elem0IsOwner = false;
-}
-
-CFaceOfElement::CFaceOfElement(const unsigned short VTK_Type, const unsigned short nPoly, const unsigned long* Nodes) {
-  /* Set the default values of the member variables. */
-  nCornerPoints = 0;
-  cornerPoints[0] = cornerPoints[1] = cornerPoints[2] = cornerPoints[3] = ULONG_MAX;
-  elemID0 = elemID1 = ULONG_MAX;
-  nPolyGrid0 = nPolyGrid1 = 0;
-  nPolySol0 = nPolySol1 = 0;
-  nDOFsElem0 = nDOFsElem1 = 0;
-  elemType0 = elemType1 = 0;
-  faceID0 = faceID1 = 0;
-  periodicIndex = periodicIndexDonor = 0;
-  faceIndicator = 0;
-
-  JacFaceIsConsideredConstant = false;
-  elem0IsOwner = false;
-
-  /* Determine the face element type and set the corner points accordingly. */
-  switch (VTK_Type) {
-    case LINE: {
-      nCornerPoints = 2;
-      cornerPoints[0] = Nodes[0];
-      cornerPoints[1] = Nodes[nPoly];
-      break;
-    }
-
-    case TRIANGLE: {
-      const unsigned short ind2 = (nPoly + 1) * (nPoly + 2) / 2 - 1;
-      nCornerPoints = 3;
-      cornerPoints[0] = Nodes[0];
-      cornerPoints[1] = Nodes[nPoly];
-      cornerPoints[2] = Nodes[ind2];
-      break;
-    }
-
-    case QUADRILATERAL: {
-      const unsigned short ind2 = nPoly * (nPoly + 1);
-      const unsigned short ind3 = (nPoly + 1) * (nPoly + 1) - 1;
-      nCornerPoints = 4;
-      cornerPoints[0] = Nodes[0];
-      cornerPoints[1] = Nodes[nPoly];
-      cornerPoints[2] = Nodes[ind2];
-      cornerPoints[3] = Nodes[ind3];
-      break;
-    }
-
-    default: {
-      ostringstream message;
-      message << "Unknown VTK surface element type, " << VTK_Type;
-      SU2_MPI::Error(message.str(), CURRENT_FUNCTION);
-    }
-  }
-}
-
-bool CFaceOfElement::operator<(const CFaceOfElement& other) const {
-  if (nCornerPoints != other.nCornerPoints) return nCornerPoints < other.nCornerPoints;
-
-  for (unsigned short i = 0; i < nCornerPoints; ++i)
-    if (cornerPoints[i] != other.cornerPoints[i]) return cornerPoints[i] < other.cornerPoints[i];
-
-  return false;
-}
-
-bool CFaceOfElement::operator==(const CFaceOfElement& other) const {
-  if (nCornerPoints != other.nCornerPoints) return false;
-  for (unsigned short i = 0; i < nCornerPoints; ++i)
-    if (cornerPoints[i] != other.cornerPoints[i]) return false;
-
-  return true;
-}
-
-void CFaceOfElement::Copy(const CFaceOfElement& other) {
-  nCornerPoints = other.nCornerPoints;
-  for (unsigned short i = 0; i < nCornerPoints; ++i) cornerPoints[i] = other.cornerPoints[i];
-  for (unsigned short i = nCornerPoints; i < 4; ++i) cornerPoints[i] = ULONG_MAX;
-
-  elemID0 = other.elemID0;
-  elemID1 = other.elemID1;
-
-  nPolyGrid0 = other.nPolyGrid0;
-  nPolyGrid1 = other.nPolyGrid1;
-  nPolySol0 = other.nPolySol0;
-  nPolySol1 = other.nPolySol1;
-
-  nDOFsElem0 = other.nDOFsElem0;
-  nDOFsElem1 = other.nDOFsElem1;
-  elemType0 = other.elemType0;
-  elemType1 = other.elemType1;
-  faceID0 = other.faceID0;
-  faceID1 = other.faceID1;
-
-  periodicIndex = other.periodicIndex;
-  periodicIndexDonor = other.periodicIndexDonor;
-  faceIndicator = other.faceIndicator;
-
-  JacFaceIsConsideredConstant = other.JacFaceIsConsideredConstant;
-  elem0IsOwner = other.elem0IsOwner;
-}
-
-void CFaceOfElement::CreateUniqueNumberingWithOrientation() {
-  /*--- Determine the element type and create the unique numbering accordingly. ---*/
-  bool swapElements = false;
-  switch (nCornerPoints) {
-    case 2: {
-      /* Element is a line. Check if the node numbering must be swapped. If so
-         also the element information must be swapped, because element 0 is to
-         the left of the face and element 1 to the right. */
-      if (cornerPoints[1] < cornerPoints[0]) {
-        swap(cornerPoints[0], cornerPoints[1]);
-        swapElements = true;
-      }
-      break;
-    }
-
-    case 3: {
-      /* Element is a triangle. The vertices are sorted in increasing order.
-         If the sequence of the new numbering is opposite to the current
-         numbering, the element information must be exchanged, because
-         element 0 is to the left of the face and element 1 to the right. */
-      unsigned long nn[] = {cornerPoints[0], cornerPoints[1], cornerPoints[2]};
-      unsigned short ind = 0;
-      if (nn[1] < nn[ind]) ind = 1;
-      if (nn[2] < nn[ind]) ind = 2;
-
-      unsigned short indm1 = ind == 0 ? 2 : ind - 1;  // Next lower index.
-      unsigned short indp1 = ind == 2 ? 0 : ind + 1;  // Next upper index.
-
-      if (nn[indp1] < nn[indm1]) {
-        /* The orientation of the triangle remains the same.
-           Store the new sorted node numbering. */
-        cornerPoints[0] = nn[ind];
-        cornerPoints[1] = nn[indp1];
-        cornerPoints[2] = nn[indm1];
-      } else {
-        /* The orientation of the triangle changes. Store the new
-           sorted node numbering and set swapElements to true. */
-        cornerPoints[0] = nn[ind];
-        cornerPoints[1] = nn[indm1];
-        cornerPoints[2] = nn[indp1];
-        swapElements = true;
-      }
-
-      break;
-    }
-
-    case 4: {
-      /* Element is a quadrilateral. The vertices are sorted in increasing order
-         under the condition neighboring vertices remain neighbors. If the
-         sequence of the new numbering is opposite to the current
-         numbering, the element information must be exchanged, because
-         element 0 is to the left of the face and element 1 to the right. */
-      unsigned long nn[] = {cornerPoints[0], cornerPoints[1], cornerPoints[2], cornerPoints[3]};
-      unsigned short ind = 0;
-      if (nn[1] < nn[ind]) ind = 1;
-      if (nn[2] < nn[ind]) ind = 2;
-      if (nn[3] < nn[ind]) ind = 3;
-
-      unsigned short indm1 = ind == 0 ? 3 : ind - 1;        // Next lower index.
-      unsigned short indp1 = ind == 3 ? 0 : ind + 1;        // Next upper index.
-      unsigned short indp2 = ind >= 2 ? ind - 2 : ind + 2;  // Opposite index.
-
-      if (nn[indp1] < nn[indm1]) {
-        /* The orientation of the quadrilateral remains the same.
-           Store the new sorted node numbering. */
-        cornerPoints[0] = nn[ind];
-        cornerPoints[1] = nn[indp1];
-        cornerPoints[2] = nn[indp2];
-        cornerPoints[3] = nn[indm1];
-      } else {
-        /* The orientation of the quadrilateral changes. Store the new
-           sorted node numbering and set swapElements to true. */
-        cornerPoints[0] = nn[ind];
-        cornerPoints[1] = nn[indm1];
-        cornerPoints[2] = nn[indp2];
-        cornerPoints[3] = nn[indp1];
-        swapElements = true;
-      }
-
-      break;
-    }
-
-    default: {
-      ostringstream message;
-      message << "Unknown surface element type with " << nCornerPoints << " corners." << endl;
-      SU2_MPI::Error(message.str(), CURRENT_FUNCTION);
-    }
-  }
-
-  /* Swap the element information, if needed. */
-  if (swapElements) {
-    swap(elemID0, elemID1);
-    swap(nPolyGrid0, nPolyGrid1);
-    swap(nPolySol0, nPolySol1);
-    swap(nDOFsElem0, nDOFsElem1);
-    swap(elemType0, elemType1);
-    swap(faceID0, faceID1);
-  }
-}
-
-void CBoundaryFace::Copy(const CBoundaryFace& other) {
-  VTK_Type = other.VTK_Type;
-  nPolyGrid = other.nPolyGrid;
-  nDOFsGrid = other.nDOFsGrid;
-  globalBoundElemID = other.globalBoundElemID;
-  domainElementID = other.domainElementID;
-  Nodes = other.Nodes;
-}
-
-CMatchingFace::CMatchingFace() {
-  nCornerPoints = 0;
-  nDim = 0;
-  nPoly = 0;
-  nDOFsElem = 0;
-  elemType = 0;
-  elemID = 0;
-
-  cornerCoor[0][0] = cornerCoor[0][1] = cornerCoor[0][2] = 0.0;
-  cornerCoor[1][0] = cornerCoor[1][1] = cornerCoor[1][2] = 0.0;
-  cornerCoor[2][0] = cornerCoor[2][1] = cornerCoor[2][2] = 0.0;
-  cornerCoor[3][0] = cornerCoor[3][1] = cornerCoor[3][2] = 0.0;
-
-  tolForMatching = 0.0;
-}
-
-bool CMatchingFace::operator<(const CMatchingFace& other) const {
-  /* First compare the number of corner points. ---*/
-  if (nCornerPoints != other.nCornerPoints) return nCornerPoints < other.nCornerPoints;
-
-  /*--- Determine the tolerance for comparing both objects. ---*/
-  const su2double tol = min(tolForMatching, other.tolForMatching);
-
-  /*--- Loop over the number of corner points and dimensions and compare the
-        coordinates. If considered different, return true if the current face
-        is considered smaller and false otherwise.            ---*/
-  for (unsigned short k = 0; k < nCornerPoints; ++k) {
-    for (unsigned short l = 0; l < nDim; ++l) {
-      if (fabs(cornerCoor[k][l] - other.cornerCoor[k][l]) > tol) return cornerCoor[k][l] < other.cornerCoor[k][l];
-    }
-  }
-
-  /*--- Both objects are considered the same. Return false. ---*/
-  return false;
-}
-
-void CMatchingFace::SortFaceCoordinates() {
-  /*--- Determine the tolerance for a matching point for this face. This is
-        accomplished by computing the minimum distance between the points of
-        the face, multiplied by a relative tolerance.        ---*/
-  for (unsigned short k = 0; k < nCornerPoints; ++k) {
-    for (unsigned short j = (k + 1); j < nCornerPoints; ++j) {
-      su2double dist = 0.0;
-      for (unsigned short l = 0; l < nDim; ++l) {
-        su2double ds = cornerCoor[k][l] - cornerCoor[j][l];
-        dist += ds * ds;
-      }
-      dist = sqrt(dist);
-
-      if (k == 0 && j == 1)
-        tolForMatching = dist;
-      else
-        tolForMatching = min(tolForMatching, dist);
-    }
-  }
-
-  tolForMatching *= 1.e-6;
-
-  /*--- Sort the points in increasing order based on the coordinates.
-        An insertion sort algorithm is used, which is quite efficient
-        for at most four corner points.                        ---*/
-  for (unsigned short k = 1; k < nCornerPoints; ++k) {
-    for (unsigned short j = k; j > 0; --j) {
-      /* Check if cornerCoor[j] is considered less than cornerCoor[j-1]. */
-      bool lessThan = false;
-      for (unsigned short l = 0; l < nDim; ++l) {
-        if (fabs(cornerCoor[j][l] - cornerCoor[j - 1][l]) > tolForMatching) {
-          lessThan = cornerCoor[j][l] < cornerCoor[j - 1][l];
-          break;
-        }
-      }
-
-      /* If cornerCoor[j] is less than cornerCoor[j-1] they must be swapped.
-         Otherwise an exit can be made from the j-loop.         */
-      if (lessThan) {
-        for (unsigned short l = 0; l < nDim; ++l) swap(cornerCoor[j][l], cornerCoor[j - 1][l]);
-      } else
-        break;
-    }
-  }
-}
-
-void CMatchingFace::Copy(const CMatchingFace& other) {
-  nCornerPoints = other.nCornerPoints;
-  nDim = other.nDim;
-  nPoly = other.nPoly;
-  nDOFsElem = other.nDOFsElem;
-  elemType = other.elemType;
-  elemID = other.elemID;
-
-  for (unsigned short k = 0; k < nCornerPoints; ++k) {
-    for (unsigned l = 0; l < nDim; ++l) {
-      cornerCoor[k][l] = other.cornerCoor[k][l];
-    }
-  }
-
-  tolForMatching = other.tolForMatching;
-}
 
 void CPhysicalGeometry::LoadLinearlyPartitionedPointsFEM(CConfig* config, CMeshReaderBase* mesh) {
   /*--- Get the partitioned coordinates and their global IDs from the mesh object. ---*/
