+++ Common/src/geometry/meson.build
@@ -1,5 +1,6 @@
 common_src += files(['CGeometry.cpp',
                      'CPhysicalGeometry.cpp',
+                     'CPhysicalGeometryFEM.cpp',
                      'CMultiGridGeometry.cpp',
                      'CDummyGeometry.cpp',
                      'CMultiGridQueue.cpp'])
