+++ Common/src/toolboxes/fem/meson.build
@@ -1,0 +1,5 @@
+common_src += files(['CFaceOfElement.cpp',
+                     'CMatchingFace.cpp',
+                     'CReorderElements.cpp',
+                     'CSortBoundaryFaces.cpp',
+                     'CSortFaces.cpp'])
